requirements.txt
backcall==0.2.0
backports.entry-points-selectable==1.1.0
colorama==0.4.4
decorator==5.1.0
distlib==0.3.3
filelock==3.3.1
importlib-metadata==4.8.1
ipython==7.28.0
jedi==0.18.0
matplotlib-inline==0.1.3
parso==0.8.2
pbr==5.6.0
pickleshare==0.7.5
pip==21.3.1
platformdirs==2.4.0
prompt-toolkit==3.0.21
Pygments==2.10.0
pyreadline==2.1
setuptools==58.3.0
six==1.16.0
stevedore==3.5.0
traitlets==5.1.0
typing-extensions==3.10.0.2
virtualenv==20.9.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
zipp==3.6.0
requests~=2.26.0
bs4~=0.0.1
beautifulsoup4~=4.10.0
Pillow~=8.4.0